Calculation time logic and show the flowers that need to be watered in wateringFragment

+++ app/src/main/java/com/example/loveplant/PlantFragment.java
@@ -8,6 +8,7 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -46,13 +47,11 @@ public class PlantFragment extends Fragment {
         recyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
         recyclerView.setAdapter(recyclerViewAdopterP);
 
-
         //swipe to remove items
         itemTouchHelper();
 
         Log.d("Sui plantrecycleView","is loading");
         return v;
-
 
     }
 
@@ -62,27 +61,45 @@ public class PlantFragment extends Fragment {
         listPlants = new ArrayList<>();
         Log.d("Sui getdata","from database");
 
-
-
         //get data from database and show them in Plant.
         List<PlantInfo>plantInfos = MainActivity.myAppDatabase.myDao().getPlantInfo();
 
         System.out.println("Size of list = " + plantInfos.size());
 
-
         for(int i = 0; i<plantInfos.size(); i++)
         {
 
             String timeStamp = plantInfos.get(i).getTimeStampe();
-            String timeNow = new SimpleDateFormat("dd-MM-yyyy").format(Calendar.getInstance().getTime());
-            dayDiff(timeNow,timeStamp,"dd-MM-yyyy");
+            String timeNow = new SimpleDateFormat("yyyyMMdd_HHmmss").format(Calendar.getInstance().getTime());
 
-            
-            int daysLeft = Integer.parseInt(plantInfos.get(i).getDay()) - dayDiff(timeNow,timeStamp,"dd-MM-yyyy HH:mm:ss");
+            Log.d("sui timeStamp"," is " + timeStamp);
+            Log.d("sui timeNow"," is " + timeNow);
+
+            Integer dayD = dayDiff(timeNow,timeStamp,"yyyyMMdd_HHmmss");
+
+            Log.d("sui dayDiff", " is " + dayD );
+
+            int daysLeft = Integer.parseInt(plantInfos.get(i).getDay()) - dayDiff(timeNow,timeStamp,"yyyyMMdd_HHmmss");
+            if(daysLeft == 0)
+            {
+
+                // update timestampe is timeNow.
+
+                //timeStamp = timeNow;
+                PlantInfo plantInfo = new PlantInfo();
+                plantInfo.setTimeStampe(timeNow);
+                MainActivity.myAppDatabase.myDao().updateTimeStamp(plantInfo);
+                Toast.makeText(getActivity(),"timeStamp updated..",Toast.LENGTH_SHORT).show();
+
+
+
+                //send this plant image to wateringfragment and show it.
+
+
+            }
+
             Log.d("Sui getdata","for loop");
-
             Log.d("sui daysLeft" ," is " +daysLeft);
-
 
             listPlants.add(new Plant(plantInfos.get(i).getImage(),plantInfos.get(i).getName(),
                     plantInfos.get(i).getDay(),String.valueOf(daysLeft)));
@@ -114,10 +131,14 @@ public class PlantFragment extends Fragment {
                 for(int i = 0 ; i < plantDbSize; i++){
 
                     PlantInfo plantInfo = MainActivity.myAppDatabase.myDao().getPlantInfo().get(i);
-                    if(plantInfo.getName().equals(name))
+                    if(plantInfo != null)
                     {
-                        MainActivity.myAppDatabase.myDao().deletePlantInfo(plantInfo);
+                        if(plantInfo.getName().equals(name))
+                        {
+                            MainActivity.myAppDatabase.myDao().deletePlantInfo(plantInfo);
+                        }
                     }
+
                 }
             }
         }).attachToRecyclerView(recyclerView);
